FIX auth, authController and auth tests
