+++ src/server/routes/auth.js
@@ -1,34 +1,13 @@
 const { UserModel } = require('../models/user')
 const jwt = require('jsonwebtoken')
 const mongoose = require('mongoose')
-const Joi = require('@hapi/joi')
 const bcrypt = require('bcrypt')
 const _ = require('lodash')
 const express = require('express')
 const router = express.Router()
+const authController = require('../controllers/authController')
 
-router.post('/', async (req, res) => {
-  const { error } = validate(req.body)
-  if (error) return res.status(400).send(error.details[0].message)
-
-  let user = await UserModel.findOne({ email: req.body.email })
-  if(!user) return res.status(400).send('Invalid email or password.')
-
-  const validPassword = await bcrypt.compare(req.body.password, user.password)
-  if (!validPassword) return res.status(400).send('Invalid email or password.')
-
-  const token = user.generateAuthToken()
-
-  res.send({token, 'id': user._id})
-})
-
-function validate(req) {
-  const schema = Joi.object({
-    email: Joi.string().min(5).max(255).required().email(),
-    password: Joi.string().min(5).max(255).required()
-  })
-  return schema.validate(req)
-}
+router.post('/', authController.login)
 
 
 module.exports = router
