+++ src/server/test/user-auth.js
@@ -287,7 +287,7 @@ if (mongoose.connection.name === 'lookahead-test') {
             })
           .end((err, res) => {
             expect(res).to.have.status(401)
-            expect(res.body.message).to.equal('Incorrect Email or Password.')
+            expect(res.body.message).to.equal('Invalid Email or Password.')
             done()
           })
         })
@@ -302,7 +302,7 @@ if (mongoose.connection.name === 'lookahead-test') {
             })
           .end((err, res) => {
             expect(res).to.have.status(401)
-            expect(res.body.message).to.equal('Incorrect Email or Password.')
+            expect(res.body.message).to.equal('Invalid Email or Password.')
             done()
           })
